+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 28/11/2019
+- Add `is_highlighted` field to dashboards
+
 # 26/11/2019
 - Add `application` field to dashboards
 - Add `application` field to topics